Fix: don't use ng to ensure proper nick completion
This is an ugly, non-ng temp fix to ensure that the DOM node value of
the command input is sent instead of the command model. When completing
a nick, there's a sequence to respect in order to update the value and
the caret position afterwards. If we update the model in there, it's
applied by ng after having updated the caret position and, since it
overwrites the DOM input value, puts the caret position at the end.