Handle keypress events as well
keydown is not enough, because browsers. A (very very very) long
explanation that I read in part is at http://unixpapa.com/js/key.html

E.g. on my keyboard layout (neo2), I don't get a keydown event for alt+<,
which in QWERTY keys is alt + caps lock + u (or the key above shift next to
enter on the right side, which actually isn't present on QWERTY (not ' but to
the right of it - it's not there on a US layout, but it is # on QWERTZ). I do get
a keypress event though.