Introduce a new media query list for wide screens
This change introduces a new media query that activates if the window is
wider than 1400px. It will then increase bufferlist width from 140px to
200px and nicklist from 100px to 140px.

If someone has better ideas than 1400px as a cutoff I am all ears to
that.

There might also be more changes we could do that further improves upon
this concept.