Document how to run everything from one port.
This commit adds some more documentation about encryption to the static landing page of Glowing Bear. It describes how to configure Apache as a server that serves the Glowing Bear static HTML content and also acts a a reverse proxy for the WeeChat WebSocket connection from a single port using HTTPS. This is an advantage for users in networks that do not allow none standard outgoing ports.