Fix Ctrl+W when the cursor follows a space.
This key binding does "Delete from cursor to previous space".
When the text before the cursor was `some example`,
it would find the space after "some" and delete "example".
When hitting Ctrl+W *again*, it would find the same space again…
and delete nothing.

This changes the code to ignore trailing spaces before the cursor
for the purspose of finding the previous space,
so that something (if at all possible) is always deleted.