allow to set timezone in settings
currently glowing-bear always show message timestamp in browser timezone.
It's common that people want to use different tz like UTC in irc.

In current code, we are using `date` filter in angular for time.
It has a 3rd `timezone` param but ignored, which will use brower timezone:

    https://docs.angularjs.org/api/ng/filter/date

This patch add a new `timezone` option in settings, and pass it to date filter.
It doesn't change/break anything when blank, but we have he ability to change timezone now.

Signed-off-by: Guo Qiao <guoqiao@gmail.com>